Move recursive endpoints into dedicated submodule (#4200)
